+++ apps/v4/app/(create)/preview/[base]/[name]/page.tsx
@@ -6,17 +6,18 @@ import { siteConfig } from "@/lib/config"
 import { absoluteUrl } from "@/lib/utils"
 import { TailwindIndicator } from "@/components/tailwind-indicator"
 import { BASES, type Base, type BaseName } from "@/registry/config"
-import { ActionMenuScript } from "@/app/(create)/components/action-menu"
-import { DesignSystemProvider } from "@/app/(create)/components/design-system-provider"
-import { HistoryScript } from "@/app/(create)/components/history-buttons"
-import { DarkModeScript } from "@/app/(create)/components/mode-switcher"
-import { PreviewStyle } from "@/app/(create)/components/preview-style"
-import { RandomizeScript } from "@/app/(create)/components/random-button"
+import { ActionMenuScript } from "@/app/(app)/create/components/action-menu"
+import { DesignSystemProvider } from "@/app/(app)/create/components/design-system-provider"
+import { HistoryScript } from "@/app/(app)/create/components/history-buttons"
+import { DarkModeScript } from "@/app/(app)/create/components/mode-switcher"
+import { OpenPresetScript } from "@/app/(app)/create/components/open-preset"
+import { PreviewStyle } from "@/app/(app)/create/components/preview-style"
+import { RandomizeScript } from "@/app/(app)/create/components/random-button"
 import {
   getBaseComponent,
   getBaseItem,
   getItemsForBase,
-} from "@/app/(create)/lib/api"
+} from "@/app/(app)/create/lib/api"
 
 export const revalidate = false
 export const dynamic = "force-static"
@@ -135,10 +136,11 @@ export default async function BlockPage({
   }
 
   return (
-    <div className="relative">
+    <div className="relative bg-background">
       <PreventScrollOnFocusScript />
       <PreviewStyle />
       <ActionMenuScript />
+      <OpenPresetScript />
       <RandomizeScript />
       <HistoryScript />
       <DarkModeScript />