+++ apps/v4/registry/bases/base/blocks/preview/cards/ui-elements.tsx
@@ -71,7 +71,7 @@ export function UIElements() {
             <Button>Button</Button>
             <Button variant="secondary">Secondary</Button>
             <Button variant="outline">Outline</Button>
-            <Button variant="destructive">Delete</Button>
+            <Button variant="ghost">Ghost</Button>
           </div>
           <Item variant="outline">
             <ItemContent>
@@ -135,8 +135,10 @@ export function UIElements() {
         <div className="flex items-center gap-4">
           <AlertDialog>
             <AlertDialogTrigger render={<Button variant="outline" />}>
-              <span className="hidden md:block">Alert Dialog</span>
-              <span className="block md:hidden">Dialog</span>
+              <span className="hidden md:flex style-sera:md:hidden">
+                Alert Dialog
+              </span>
+              <span className="flex md:hidden style-sera:md:flex">Dialog</span>
             </AlertDialogTrigger>
             <AlertDialogContent size="sm">
               <AlertDialogHeader>
@@ -153,7 +155,10 @@ export function UIElements() {
             </AlertDialogContent>
           </AlertDialog>
           <ButtonGroup>
-            <Button variant="outline">Button Group</Button>
+            <Button variant="outline">
+              <span className="style-sera:hidden">Button Group</span>
+              <span className="hidden style-sera:block">Group</span>
+            </Button>
             <DropdownMenu>
               <DropdownMenuTrigger
                 render={<Button variant="outline" size="icon" />}