+++ apps/v4/registry/bases/radix/blocks/preview/cards/ui-elements.tsx
@@ -71,7 +71,7 @@ export function UIElements() {
             <Button>Button</Button>
             <Button variant="secondary">Secondary</Button>
             <Button variant="outline">Outline</Button>
-            <Button variant="destructive">Delete</Button>
+            <Button variant="ghost">Ghost</Button>
           </div>
           <Item variant="outline">
             <ItemContent>
@@ -136,8 +136,12 @@ export function UIElements() {
           <AlertDialog>
             <AlertDialogTrigger asChild>
               <Button variant="outline">
-                <span className="hidden md:block">Alert Dialog</span>
-                <span className="block md:hidden">Dialog</span>
+                <span className="hidden md:flex style-sera:md:hidden">
+                  Alert Dialog
+                </span>
+                <span className="flex md:hidden style-sera:md:flex">
+                  Dialog
+                </span>
               </Button>
             </AlertDialogTrigger>
             <AlertDialogContent size="sm">
@@ -155,7 +159,10 @@ export function UIElements() {
             </AlertDialogContent>
           </AlertDialog>
           <ButtonGroup>
-            <Button variant="outline">Button Group</Button>
+            <Button variant="outline">
+              <span className="style-sera:hidden">Button Group</span>
+              <span className="hidden style-sera:block">Group</span>
+            </Button>
             <DropdownMenu>
               <DropdownMenuTrigger asChild>
                 <Button variant="outline" size="icon">